feat(console): JITSU-213 read-only get_audit_log MCP tool
Extract the audit-log read path (access checks, secret masking, prev/next
→ diff reduction, origin filter, cursor paging) from pages/api/audit-log.ts
into lib/server/audit-log-service.ts and expose it as the MCP tool
`get_audit_log`, so an agent can read back what changed and by whom —
including changes made through MCP itself (origin "mcp"). The HTTP route
is now a thin wrapper over the same service.

Behavior change: unknown values in the `origin` filter now return 400
instead of being silently ignored.
